+++ src/pages/kakaoLogin/Kakao.jsx
@@ -1,9 +1,32 @@
 //Kakao.js
 import React from "react";
+import { useNavigate } from "react-router-dom";
+import { useEffect } from "react";
+import axios from "axios";
 import { useCookies } from "react-cookie";
+import Home from "../Home";
 
 const Kakao = (props) => {
   const [cookie, setCookie, removeCookie] = useCookies(["authorization"]);
+
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    // 현재 URL에서 쿼리 파라미터 추출
+    const urlParams = new URLSearchParams(window.location.search);
+    const token = urlParams.get("token");
+    console.log("urlparan", urlParams);
+    console.log("token ", token);
+
+    if (token) {
+      localStorage.setItem("authorization", JSON.stringify(`Bearer ${token}`));
+
+      setCookie("authorization", `Bearer ${token}`, { path: "/" });
+      console.log("token");
+      navigate("/");
+    }
+    // 추출한 토큰을 로컬 스토리지에 저장
+  }, [navigate]);
 
   return (
     <div>
