+++ src/pages/kakaoLogin/SocialKakao.jsx
@@ -13,21 +13,21 @@ const SocialKakao = () => {
 
   const navigate = useNavigate();
 
-  useEffect(() => {
-    // 현재 URL에서 쿼리 파라미터 추출
-    const urlParams = new URLSearchParams(window.location.search);
-    const token = urlParams.get("token");
-    console.log("urlparan", urlParams);
-    console.log("token ", token);
+  // useEffect(() => {
+  //   // 현재 URL에서 쿼리 파라미터 추출
+  //   const urlParams = new URLSearchParams(window.location.search);
+  //   const token = urlParams.get("token");
+  //   console.log("urlparan", urlParams);
+  //   console.log("token ", token);
 
-    if (token) {
-      localStorage.setItem("authorization", JSON.stringify(`Bearer ${token}`));
-      setCookie("authorization", `Bearer ${token}`);
-      console.log("token");
-      navigate("/");
-    }
-    // 추출한 토큰을 로컬 스토리지에 저장
-  }, [navigate]);
+  //   if (token) {
+  //     localStorage.setItem("authorization", JSON.stringify(`Bearer ${token}`));
+  //     setCookie("authorization", `Bearer ${token}`);
+  //     console.log("token");
+  //     navigate("/");
+  //   }
+  //   // 추출한 토큰을 로컬 스토리지에 저장
+  // }, [navigate]);
 
   const { search } = useLocation();
   const token = queryString.parse(search);
